Refactor app into modular layout and utilities
